+++ src/GRA.Controllers/ViewModel/MissionControl/Participants/LogActivityViewModel.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc.Rendering;
 using System.ComponentModel;
+using GRA.Domain.Model;
 
 namespace GRA.Controllers.ViewModel.MissionControl.Participants
 {
@@ -10,12 +11,14 @@ namespace GRA.Controllers.ViewModel.MissionControl.Participants
         public int HouseholdCount { get; set; }
         public int PrizeCount { get; set; }
         public bool HasAccount { get; set; }
-        [DisplayName("Minutes Read")]
-        public int? MinutesRead { get; set; }
+        [DisplayName("Activity Amount")]
+        public int? ActivityAmount { get; set; }
+        public bool ShowSecretCode { get; set; }
         [DisplayName("Secret Code")]
         public string SecretCode { get; set; }
         public bool IsSecretCode { get; set; }
         public int? VendorCodeTypeId { get; set; }
         public SelectList VendorCodeTypeList { get; set; }
+        public PointTranslation PointTranslation { get; set; }
     }
 }